requirements.txt
Django==3.2.9
django-allauth==0.46.0
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-select2==7.9.0
django-simple-captcha==0.5.14
django_tables2==2.4.1
django-tawkto==0.3
django-widget-tweaks==1.4.9
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.2
reportlab==3.6.3
whitenoise==5.3.0
